Changed wrong package names (now all the names are written in small letters). Fixed GetListOfOrdersTest and OrderParameterizedColorTest (added checking status code and dody in every test). Added additional information on parameterized test. Added Lombok annotations. Delete GSON from ScooterAPI. Added annotation @JsonInclude(Include.NON_NULL) into needed dto classes.
