--- src/main/java/Scooter/DTO/IDCourier.java
@@ -1,18 +1,0 @@
-package Scooter.DTO;
-
-public class IDCourier {
-    private int ID;
-
-    public IDCourier(int ID) {
-        this.ID = ID;
-    }
-
-    public IDCourier() {};
-    public int getID() {
-        return ID;
-    }
-
-    public void setID(int ID) {
-        this.ID = ID;
-    }
-}
