--- src/main/java/Scooter/DTO/LoginCourier.java
@@ -1,24 +1,0 @@
-package Scooter.DTO;
-
-public class LoginCourier {
-    private String login;
-    private String password;
-    public LoginCourier(String login, String password) {
-        this.login = login;
-        this.password = password;
-    }
-    public LoginCourier(){};
-    public String getLogin() {
-        return login;
-    }
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
