--- src/main/java/Scooter/DTO/NewCourier.java
@@ -1,35 +1,0 @@
-package Scooter.DTO;
-
-public class NewCourier {
-    private String login;
-    private String password;
-    private String firstName;
-
-    public NewCourier(String login, String password, String firstName) {
-        this.login = login;
-        this.password = password;
-        this.firstName = firstName;
-    }
-    public NewCourier(){};
-
-    public String getLogin() {
-        return login;
-    }
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-}
