+++ src/main/java/scooter/dto/IDCourier.java
@@ -1,0 +1,12 @@
+package scooter.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class IDCourier {
+    private int ID;
+}
