+++ src/main/java/scooter/dto/LoginCourier.java
@@ -1,0 +1,15 @@
+package scooter.dto;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginCourier {
+    private String login;
+    private String password;
+}
