+++ src/main/java/scooter/dto/NewCourier.java
@@ -1,0 +1,15 @@
+package scooter.dto;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class NewCourier {
+    private String login;
+    private String password;
+    private String firstName;
+}
