+++ src/main/java/scooter/generatedata/courierfactory/LoginFactory.java
@@ -1,4 +1,4 @@
-package Scooter.GenerateData.CourierFactory;
+package scooter.generatedata.courierfactory;
 
 import io.qameta.allure.Step;
 import org.apache.commons.lang3.RandomStringUtils;
